+++ UserManagement/Calculator/Modtion.cs
@@ -8,7 +8,7 @@ namespace UserManagement.Calculator
 {
     public class Modtion
     {
-        public static float Mod(int NumberOne, int NumberTwo)
+        public static float Mod(float NumberOne, float NumberTwo)
         {
             float mod = 0;
             mod = NumberOne % NumberTwo;
